+++ elfinder_pyramid_example/elfinder_pyramid_example/static/elfinder/jquery/ui-themes/smoothness/jquery-ui-1.10.1.custom.css
